[refactor] 상수 처리
- 숫자 및 에러 메세지 등 Constant에 상수 처리
