--- src/main/java/racingcar/Constant.java
@@ -1,5 +1,0 @@
-package racingcar;
-
-public class Constant {
-    public static final String MOVE_MARK = "-";
-}
