+++ src/main/java/racingcar/util/Constant.java
@@ -1,0 +1,11 @@
+package racingcar.util;
+
+public class Constant {
+    public static final String MOVE_MARK = "-";
+
+    public static final Integer RANDOM_MAXIMUM_NUMBER = 9;
+    public static final Integer FORWARD_STANDARD_NUMBER = 4;
+
+    public static final String CAR_NAME_LENGTH_ERROR_MESSAGE = "[ERROR] 자동차 이름은 5자 이하만 가능하다.";
+    public static final String NOT_DIGIT_ERROR_MESSAGE = "[ERROR] 시도 횟수는 숫자여야 한다.";
+}
